+++ app/src/main/java/com/example/cookit/MainActivity.kt
@@ -8,6 +8,10 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.cookit.screens.createRecipe.CreateRecipeScreen
+import com.example.cookit.screens.home.HomeScreen
+import com.example.cookit.screens.logIn.LogInScreen
+import com.example.cookit.screens.signUp.SignUpScreen
 import com.example.cookit.ui.theme.CookItTheme
 
 class MainActivity : ComponentActivity() {
@@ -30,10 +34,9 @@ fun App() {
         navController = navController,
         startDestination = "logIn"
     ) {
-        composable("logIn") {LogInScreen(navController)}
-        composable("signUp") {SignUpScreen(navController)}
-        composable("home") {HomeScreen(navController)}
+        composable("logIn") { LogInScreen(navController) }
+        composable("signUp") { SignUpScreen(navController) }
+        composable("home") { HomeScreen(navController) }
         composable("createRecipe") { CreateRecipeScreen(navController) }
-        //composable("home") {HomeScreen(navController)}
     }
 }
