+++ app/src/main/java/com/example/cookit/screens/logIn/LogInScreen.kt
@@ -1,5 +1,4 @@
-package com.example.cookit
-
+package com.example.cookit.screens.logIn
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -8,72 +7,54 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.runtime.*
 import androidx.navigation.NavHostController
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.Alignment
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Lock
+import androidx.compose.material.icons.filled.Person
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.sp
+import com.example.cookit.R
 import com.google.firebase.auth.FirebaseAuth
 
-
 @Composable
-fun SignUpScreen(navController: NavHostController) {
+fun LogInScreen(navController: NavHostController) {
     val auth = FirebaseAuth.getInstance()
+
+    var username by remember { mutableStateOf(TextFieldValue("")) }
+    var password by remember { mutableStateOf(TextFieldValue("")) }
+    var passwordError by remember { mutableStateOf("") }
+    var loginError by remember { mutableStateOf("") }
+    var isLoading by remember { mutableStateOf(false) }
+
+    // Handle the login logic
+    fun handleLogin() {
+        isLoading = true
+        loginError = "" // Clear previous errors
+        auth.signInWithEmailAndPassword(username.text, password.text)
+            .addOnCompleteListener { task ->
+                isLoading = false
+                if (task.isSuccessful) {
+                    // Navigate to the home screen upon successful login
+                    navController.navigate("home")
+                } else {
+                    // Show error message
+                    loginError = task.exception?.localizedMessage ?: "Login failed"
+                }
+            }
+    }
+
+    // Get screen width and height
     val configuration = LocalConfiguration.current
     val screenWidth = configuration.screenWidthDp.dp
     val screenHeight = configuration.screenHeightDp.dp
-
-    var username by remember { mutableStateOf(TextFieldValue("")) }
-    var email by remember { mutableStateOf(TextFieldValue("")) }
-    var password by remember { mutableStateOf(TextFieldValue("")) }
-    var passwordError by remember { mutableStateOf("") }
-    var signUpError by remember { mutableStateOf("") }
-    var isLoading by remember { mutableStateOf(false) }
-
-    fun validatePassword(username: String, password: String): Boolean {
-        return when {
-            password.length < 8 -> {
-                passwordError = "Password must be at least 8 characters"
-                false
-            }
-            password == username -> {
-                passwordError = "Password cannot be the username"
-                false
-            }
-            else -> {
-                passwordError = ""
-                true
-            }
-        }
-    }
-
-    fun handleSignUp() {
-        if (validatePassword(username.text, password.text)) {
-            isLoading = true
-            signUpError = "" // Clear previous errors
-            auth.createUserWithEmailAndPassword(username.text, password.text)
-                .addOnCompleteListener { task ->
-                    isLoading = false
-                    if (task.isSuccessful) {
-                        // Navigate to home screen after successful sign-up
-                        navController.navigate("home")
-                    } else {
-                        // Display the error message
-                        signUpError = task.exception?.localizedMessage ?: "Sign up failed"
-                    }
-                }
-        }
-    }
 
     Scaffold(
         modifier = Modifier.fillMaxSize(),
@@ -95,44 +76,29 @@ fun SignUpScreen(navController: NavHostController) {
                     horizontalAlignment = Alignment.CenterHorizontally
                 ) {
 
-                    Spacer(modifier = Modifier.height(screenHeight * 0.1f))
+                    Spacer(modifier = Modifier.height(50.dp))
 
                     Image(
                         painter = painterResource(id = R.drawable.logo),
                         contentDescription = "App Logo",
                         modifier = Modifier
-                            .fillMaxWidth(0.6f)
-                            .aspectRatio(1f)
+                            .fillMaxWidth()
+                            .height(screenWidth * 0.6f)
+                            .padding(top = 16.dp)
                     )
 
-                    Spacer(modifier = Modifier.height(screenHeight * 0.05f))
-
-                    Text(
-                        text = "Just a few quick things to get started",
-                        fontSize = (screenWidth.value * 0.05f).sp,
-                        modifier = Modifier.padding(bottom = 16.dp)
-                    )
+                    Spacer(modifier = Modifier.height(50.dp))
 
                     OutlinedTextField(
                         value = username,
                         onValueChange = { username = it },
-                        placeholder = { Text("john.doe") },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(vertical = 8.dp),
-                        singleLine = true,
-                        shape = RoundedCornerShape(25.dp),
-                    )
-
-                    OutlinedTextField(
-                        value = email,
-                        onValueChange = { email = it },
                         placeholder = { Text("john.doe@example.com") },
                         modifier = Modifier
                             .fillMaxWidth()
                             .padding(vertical = 8.dp),
                         singleLine = true,
                         shape = RoundedCornerShape(25.dp),
+                        leadingIcon = { Icon(Icons.Default.Person, contentDescription = "Username Icon") }
                     )
 
                     OutlinedTextField(
@@ -144,31 +110,22 @@ fun SignUpScreen(navController: NavHostController) {
                             .padding(vertical = 8.dp),
                         singleLine = true,
                         shape = RoundedCornerShape(25.dp),
+                        leadingIcon = { Icon(Icons.Default.Lock, contentDescription = "Password Icon") }
                     )
 
-                    if (passwordError.isNotEmpty()) {
+                    if (loginError.isNotEmpty()) {
                         Text(
-                            text = passwordError,
+                            text = loginError,
                             color = Color.Red,
-                            fontSize = (screenWidth.value * 0.04f).sp,
-                            modifier = Modifier.padding(vertical = 4.dp)
-                        )
-                    }
-
-                    if (signUpError.isNotEmpty()) {
-                        Text(
-                            text = signUpError,
-                            color = Color.Red,
-                            fontSize = (screenWidth.value * 0.04f).sp,
-                            modifier = Modifier.padding(vertical = 4.dp)
+                            modifier = Modifier.padding(vertical = 8.dp)
                         )
                     }
 
                     FilledTonalButton(
-                        onClick = { handleSignUp() },
+                        onClick = { handleLogin() },
                         modifier = Modifier
                             .fillMaxWidth()
-                            .height(screenHeight * 0.07f),
+                            .padding(vertical = 8.dp),
                         enabled = !isLoading
                     ) {
                         if (isLoading) {
@@ -177,11 +134,10 @@ fun SignUpScreen(navController: NavHostController) {
                                 modifier = Modifier.size(16.dp)
                             )
                         } else {
-                            Text("Create Account")
+                            Text("Log In")
                         }
                     }
 
-                    Spacer(modifier = Modifier.height(screenHeight * 0.05f))
 
                     Column(
                         modifier = Modifier
@@ -195,7 +151,7 @@ fun SignUpScreen(navController: NavHostController) {
                             verticalAlignment = Alignment.CenterVertically,
                         ) {
                             Text(
-                                text = "Already have an account?",
+                                text = "Don't have an account?",
                                 color = Color.Black,
                                 fontSize = 16.sp,
                                 modifier = Modifier.padding(end = 8.dp)
@@ -203,8 +159,7 @@ fun SignUpScreen(navController: NavHostController) {
 
                             Surface(
                                 onClick = {
-                                    //PRESSABLE FUNCTIONALITY HAS TO CHANGE TO HOME SCREEN
-                                    navController.navigate("logIn")
+                                    navController.navigate("signUp")
                                 },
                                 modifier = Modifier
                                     .height(48.dp)
@@ -214,7 +169,7 @@ fun SignUpScreen(navController: NavHostController) {
                             ) {
                                 Box(contentAlignment = Alignment.Center) {
                                     Text(
-                                        text = "Log In",
+                                        text = "Sign up",
                                         modifier = Modifier.padding(8.dp),
                                         color = Color.Blue,
                                         fontSize = 16.sp
@@ -228,4 +183,6 @@ fun SignUpScreen(navController: NavHostController) {
         }
     )
 }
+
+
